some modifications

+++ code2048/v4_2048_32.c
@@ -5,9 +5,9 @@
 #include<time.h>
 
 int N = 2048;
-double A[2048][2048];
-double array1[2048][2048];
-double array2[2048][2048];
+int A[2048][2048];
+int array1[2048][2048];
+int array2[2048][2048];
 int B = 512;
 
 void init(){
