+++ matmul.c
@@ -5,9 +5,9 @@
 #include<time.h>
 
 int N = 4;
-double A[4][4];
-double array1[4][4];
-double array2[4][4];
+int A[4][4];
+int array1[4][4];
+int array2[4][4];
 int B = 2;
 
 void init(){
